Blog: Change url structure for blog articles to be more SEO and english friendly
Changed the blog/{id} routing to blog/{year}/{month}/{day}/{fileKey} routing where fileKey corresponds to the filename of the entry (for now). This should be more SEO friendly from my understanding but will regardless make it nicer to read as a human rather than using the ID. I kept the ID route by moving it to blog/id/{id} for the controller as well as adding logic to query based on id in the compoment too.