Blog: Add support for assets keyed to blog article
To support arbitraty blog assets, aka images for now, IBlogRepository and BlogController was extended to support serving static files correlated to a blog article. This includes restructuring the format for blog posts; a single folder for an article containing a minimum of a .json and .content file sharing the same name as the folder. Any assets related to the article can then be stored in the folder and will be available to be served from the service and controller via GET /api/blog/{year}/{month}/{day}/{articleKey}/{assetKey} where assetKey is the filename of the asset.

Very naive implementation for now, will extend as needed when requried.