AboutMe: Added skills chart using ChartJs.
Probably will have to remove ChartJs and use another chart provider as there is an issue where charts in hidden divs will not render when made unhidden unless the data is updated. While I can do this it is more effort and a workaround that I'd prefer not to do with such basic charting usages.